builder updates
added helipad that can be placed with builder box. costs 1500, can only be placed near actual markers. allows garaging and ungaraging of helis only
HQ detection radius default now changes with how built up HQ is, from 500 to 1000 as you move from 0 to 2500 in defenses
time to remove building is now a third of the time to build (fourth of old time), instead of a flat 10 seconds
time to construct a building is 25% less, and decreased by an additional 25% near HQ